Mission10 Submission
In addition to the submission, the following Bootstrap features were added:

Added to the Books.razor and OrderTable.razor files: <thead class="thead-light”> which provides a light background color to the head of the table.

Also added table-hover Bootstrap class to the tables in the Books.razor and the OrderTable.razor files. This helps the user see what item they are hovering over.